Changed how we draw the world.
There is a array (chunkArray) that has the entire world preloaded and
then we only draw a small portion of it at a time. The portions we draw
is loaded into displayArray and when we draw it, we just look at each
chunk in the displayArray and call its DisplayList. IMPORTANT: There is
current nothing to catch the edge case of going outside of the world.
Right now, the code is setup to generate a 20x20 world and the player
spawns in the center at [10][10]. If you move to far you'll get an
array out of bounds error. Will fix this soon.