Please read extended description.
I changed adding blocks again. But this time what I did was I added an
array called cubeExists. This functions almost the same as worldArray,
the exception is that instead of holding type, it holds a 1 or a 0
depending on if there exists a block in that location. The reason why I
did this was because, currently if a block is obscured on all sides,
worldArray treats it as if there is no block at that location. There
isn't a huge problem with this, however there is an interesting thing
you can do with spawning blocks if we kept it this way. When I spawn a
block, I check the worldArray to see if there is a block in that
location and if there isn't, I spawn a block there. However, since the
system treats an obscured block and a location without a block the same,
you can technically spawn a block in a location that has an obscured
block. By using cubeExists, I can now tell the system that even though
worldArray sets a block type as zero and there by doesn't draw anything
at that location, I can still determine if there is suppose to be a
block in that location. By implementing this, you are now not able to
spawn a block in a position where a block exists but is obscured. Also,
you cannot spawn a block that is outside the chunk space (IE outside the
16x16x128 space).