Sounds READ EXTENDED DESCRIPTION
IMPORTANT: GO TO:

http://www.lwjgl.org/wiki/index.php?title=Slick-Util_Library_-_Introduction
And download the library. Make sure you download the whole library and
not just the single jar. There is a link to the website on the link I
gave you with a zip file download.
The jars you need to add:
slick-util.jar for textures.
jorbis-*.jar AND jogg-*.jar for sounds.
Add them the same way as you did lwjgl.
What else I've added. Now you can choose what type of block you can to
spawn. Right now I've only implemented three different types. The
default spawn is a dirt block but you can also spawn a grass and wood
block. To choose what block to spawn, you have to use the number keys. 1
= Grass, 2 = Dirt, 3 = Wood. I'll add more as I got but I think so far I
should commit what I've done so far. Also, the sounds will play
depending on what block you spawn or destroy. The default song, if there
is no sound for that specific type of block, will be the sound for a
dirt block. The minecraft files has 4 different sounds for each type of
block so what I've done is load the 4 sounds for each block into their
own array and I pick an random sound from that set of 4 each time so we
get some variety.