Cleaned up code: Extended Description
I cleaned up some of the sound code, no skyboxes yet. I also made it so
that you cannot spawn a block if there is no block touching the location
you want to spawn it at so thereby not allowing you to spawn a block in
mid air. There is a weird bug w/ spawning and deleting and it happens
when you're standing in one chunk and try to delete/spawn a cube that in
inside another chunk. For deleting, it won't delete unless I'm in the
same chunk the block is and for spawning, I can spawn a cube and it will
exist outside the current chunk but technically its inside the current
chunk. What I mean is, for instance I'm in chunk 3 and x position 0 and
I want to spawn a block in chunk 2 at x position 15. However, while
visually it looks like it spawns there, internally it technically spawns
it in chunk 3 at position 15. I'll try to see if I can fix this. Also,
changed trees so they're made of wood and not dirt.